+++ lib/src/main/java/org/asamk/signal/manager/storage/identities/TrustNewIdentity.java
@@ -1,0 +1,7 @@
+package org.asamk.signal.manager.storage.identities;
+
+public enum TrustNewIdentity {
+    ALWAYS,
+    ON_FIRST_USE,
+    NEVER
+}