+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Removed deprecated fallback data paths, only `$XDG_DATA_HOME/signal-cli` is used now
   For those still using the old paths (`$HOME/.config/signal`, `$HOME/.config/textsecure`) you need to move those to the new location.
 
+### Added
+- New global parameter `--trust-new-identities=always` to allow trusting any new identity key without verification
+
 ## [0.8.5] - 2021-08-07
 ### Added
 - Source name is included in JSON receive output (Thanks @technillogue)